big.text now accepts str/bytes subclasses too.
Previously the functions in big.text only accepted
objects of type str or bytes.  I'm fixing them to
also accept objects of a subclass of str or bytes.

If you pass in a subclass of str or bytes, internally
the function will cast it to a real str/bytes instance
before operating on it, and it will always return
str or bytes objects.