0.9!  Fixed 13-year old bug in BoundInnerClasses.
Bugfix!  If an outer class Outer had an inner class Inner
decorated with @BoundInnerClass, and o is an instance of
Outer, and o evaluated to false in a boolean context,
o.Inner would be the *unbound* version of Inner.  Now
it's the bound version, as is proper.

Modified `tests/test_boundinnerclasses.py`:

  - Added regression test for the above bugfix (of course!).
  - It now takes advantage of that newfangled "zero-argument super".
  - Added testing of an unbound subclass of an unbound subclass.