toy_multisplit is now *exactly* like multisplit.
Specifically, the result is identical to
    multisplit(keep=ALTERNATING, separate=True).
As in, it returns the empty strings too.

I also changed toy_multisplit_original to match.
And, toy_multisplit_original now also supports bytes.

Rewrote the tests for toy_multisplit* so everybody
gets tested with str and bytes.  And of course I
print out the times for everybody, formatted nicely.