Tiny optimization for toy_multisplit.
toy_multisplit sorts the separators into buckets by length.
It then progressively shaves off substrings from the front of
the string, and checks to see if a substring of each possible
length matches any of the separators.

The way we build the list of separators by length, we find the
longest separator, then pre-create places to store all the
separators, with their length.

The optimization: if you don't have any separators of one
of the lengths less than the max length, throw away the bucket.
That'll save you from testing against an empty list of separators.

e.g. if your list of separators is ('X', 'XYZ'), you don't have
any separators of length 2, so, don't bother testing to see if
s[:2] matches any separators.