Slight performance upgrade for `StateMachine`.
`StateMachine` always uses a *copy* of the observer list
(specifically, a tuple) when calling the observers; this
means it's safe to modify the observer list at any time.

`StateMachine` used to always make a fresh copy every time you
called an event; now it uses a cached copy, and only recomputes
the tuple when the observer list changes.

(Well... it's not safe to modify the observer list from two
threads simultaneously while also dispatching events.  As with
many libraries, the `StateMachine` API leaves locking up to you.)