get_int_or_float converts floats to equival-ints.
Small tweak to `get_int_or_float`.  If the `o` you pass in
is a `float`, or can be converted to `float`, `get_int_or_float`
will experimentally convert it to `int`.  If the `float` and
`int` versions compare equal, it will return the `int`, otherwise
it'll return the `float`.

For example, `get_int_or_float("13.5")` will still return `13.5`
(a `float`), but `get_int_or_float("13.0")` will now return `13`
(an `int`).