Cache for line comments, fix lines_strip.
_delimiters_to_state_machine computes the state machine states
based on the incoming delimiters.  It now uses functools.lru_cache
to memoize.  This requires split_delimiters to convert delimiters
from a dict to a tuple of 2-tuples (tuple(delimiters.items())).

If lines_strip got a line that only contained whitespace (or
strings from "separators"), it used to move the entire line into
info.leading.  That's wrong, it should go into info.trailing,
in part so we don't awkwardly change the column_number.