Fixed gnarly bug with split_delimiters!
Now that we support multicharacter delimiters and escape
strings with split_delimiters, we need to be careful.
The current implementation uses multisplit, giving it a
list of all delimiters.  But what if there are overlapping
delimiters?  What if some delimiters are prefixes of other
delimiters?

I now seem to handle these funny cases.  The basic approach
is to recognize when the situation happens, handle it
appropriately, clipping off just the text we want, then
recreate the multisplit iterator picking up where we left off.

See the split_delimiters "torture tests" in test/test_text.py.