Add file to fix tests.
I added a new directory to the test suite, but didn't add a file in it.
Git doesn't actually track directories; they are a side-effect of files.
If you have an empty directory you can't check it in to Git.
So, added a file into the test directory so Git will install it.