String.line etc is in!  Also bugfixes.
If s is a String containing newlines, and x is a slice
taken from the middle of s, then

* x.line is the entire contents of the line that x is a
  slice out of.  The previous character to x[0] is a
  linebreak (or x[0] is s[0]), and the next character
  after x[-1] is a linebreak (or x[-1] is s[-1]).
* x.previous_line is the line before x.line.
* x.next_line is the line after x.line.

These rules hold true even if x contains a newline.
If x straddles lines 3 and 4, then x.line is s lines 3 and 4,
x.previous_line is s line 2, and x.next_line is s line 5.

Also, fixed two regressions.  In both cases, we have
a String s which ends with a linebreak character,
and l = len(s):
* iter(s) wouldn't yield that last character.
* s[l:l] would advance the .line_number by 1.
  s[l:l][l:l] would advance the .line_number by 2. etc.