Fix coverage testing for Github action.
I manually run Coverage for all the supported versions of Python.
But my Github action only runs coverage on the latest version.  So
it's a constant nosebleed of "oh yeah that test code only kicks in
for the older versions" things.  (I just mark it as no cover, AFAIK
there's no way to tell Coverage "this only runs in versions X-Y").