LinkedList has defeated slice clamping.
LinkedList and list now produce identical results when slicing,
even with wild values that need to be clamped.  I think the
list clamping behavior is idiotic, but this is the price you pay
when you want to make something Pythonic--you need to feel like
Python even when it's being dumb.