Discarded linked list iterator "sloppy" semantics.
If IT is a linked list iterator, and IT pointed
at either "head" or "tail, it used to be true that
it.append and it.prepend were identical, and
it.extend and it.extendleft were also identical.
This was called "sloppy" semantics.  In truth,
you should only append to "head", and prepend to
"tail"--it seemed nonsensical to prepend to "head".

I finally figured out what the semantics should be!
If you run this:
    t = linked_list((1, 2, 3))
    it = t.find(3).after()
    it.append(4)
linked_list creates a "deleted" node after 3,
points "it" to the deleted node, and inserts
a new node for 4 between the "deleted" node
and "tail".

Symmetric for prepend and "head".
Analogously for extend and extendleft.
For reverse iterators, swap prepend and append,
  swap extend and extendleft.