Added to_{head_tail} and is_forward/is_reversed.
For linked list iterators, it.to_head() jumps the
iterator to head, and it.to_tail jumps the iterator
to tail.  (These are not reversed for reverse
iterators; they are navigation by absolute reckoning.)

Also, is_forward and is_reverse are properties returning
whether or not this iterator is a forward or reverse
iterator respectively.