If i points to a special node, i[0:1] raises.
If i is an iterator (forward or backwards) pointing at a
special node, i[0] raises a SpecialNodeError.  But in that
scenario, *slices* that access index 0 used to not raise
and instead just skipped over index 0.  (If i points to
a data node, i[-1:2] returns a linked list containing
three values; if i pointed to a special node, i[-1:2]
returned a list only containing *two* values.)

But that's a bad idea; slices that access i[0] should
raise if i points to a speical node, just like accessing
i[0] does.